(PC-28273)[PRO] refactor: OpeningHoursForm
